integral_calc_bot/requirements.txt
python-telegram-bot==20.6
sympy
matplotlib
python-dotenv